Internat doesn't depend on wxCore, ComponentInterface, MemoryX
