+++ src/ShuttleGui.h
@@ -746,4 +746,13 @@ public:
    teShuttleMode GetMode() { return  mShuttleMode; };
 };
 
+class ComponentInterfaceSymbol;
+
+//! Convenience function often useful when adding choice controls
+AUDACITY_DLL_API TranslatableStrings Msgids(
+   const EnumValueSymbol strings[], size_t nStrings);
+
+//! Convenience function often useful when adding choice controls
+AUDACITY_DLL_API TranslatableStrings Msgids( const std::vector<EnumValueSymbol> &strings );
+
 #endif
